move colormaps to new file
