+++ CMakeLists.txt
@@ -34,6 +34,7 @@ add_library(
   src/bounding_box_extraction.cpp
   src/bounding_box.cpp
   src/color.cpp
+  src/colormaps.cpp
   src/dynamic_scene_graph_layer.cpp
   src/dynamic_scene_graph.cpp
   src/edge_attributes.cpp
