+++ include/spark_dsg/color.h
@@ -37,7 +37,6 @@
 #include <array>
 #include <cstdint>
 #include <ostream>
-#include <vector>
 
 namespace spark_dsg {
 
@@ -107,76 +106,5 @@ struct Color {
   static Color fromHLS(float hue, float luminance, float saturation);
   std::array<float, 3> toHLS() const;
 };
-
-namespace colormaps {
-
-/**
- * @brief Generate a gray-scale color.
- * @param value The value of the gray [0,1, where 0 is black and 1 is white].
- */
-Color gray(float value = 0.5f);
-
-/**
- * @brief Generate a color based on a quality value.
- * @param value The quality value [0,1], where 0 is red, 0.5 is yellow, and 1 is
- * green.
- */
-Color quality(float value);
-
-/**
- * @brief Generate a color based on a spectrum of colors.
- * @param value The spectrum value [0,1], where 0 is the first color and 1 is the
- * last color.
- * @param colors The list of colors in the spectrum. Colors are assumed equidistant.
- */
-Color spectrum(float value, const std::vector<Color>& colors);
-
-/**
- * @brief Interpolate between two colors in HLS space
- * @param value A weight in [0,1] where 0 is the first color and 1 is the last
- * @param start The color corresponding to 0
- * @param end The color corresponding to 1
- */
-Color hls(float value, const Color& start, const Color& end);
-
-/**
- * @brief Interpolate between two colors in HLS space
- * @param value A weight in [0,1] where 0 is the first color and 1 is the last
- * @param start The HLS color corresponding to 0
- * @param end The HLS color corresponding to 1
- */
-Color hls(float value,
-          const std::array<float, 3>& start,
-          const std::array<float, 3>& end);
-
-/**
- * @brief Generate a color based on a ironbow value.
- * @param value The temperature value [0,1], where 0 is dark and 1 is light.
- */
-Color ironbow(float value);
-
-/**
- * @brief Generate a color based on a rainbow value.
- * @param value The rainbow value [0,1], where 0 is red, 0.5 is green, and 1 is
- * blue.
- */
-Color rainbow(float value);
-
-/**
- * @brief Generate a sequence of never repeating colors in the rainbow spectrum.
- * @param id The id of the color in the sequence.
- * @param ids_per_revolution The number of colors per revolution of the hue.
- */
-Color rainbowId(size_t id, size_t ids_per_revolution = 16);
-
-/**
- * @brief Pick a color from
- * https://colorbrewer2.org/#type=qualitative&scheme=Paired&n=12
- * @param id The id of the color in the sequence
- * @note Indices outside of the 0-11 range will get remapped to be inside the range
- */
-Color colorbrewer(size_t id);
-
-}  // namespace colormaps
 
 }  // namespace spark_dsg
